gilang

+++ application/modules/sharepricesname/controllers/RequestController.php
@@ -183,8 +183,11 @@ class Sharepricesname_RequestController extends Zend_Controller_Action
 		);
 		try {
 			$_q = $this->_modelSp->select()
-				->where('SHAREPRICES_NAME = ?', $this->_posts['SHAREPRICES_NAME']);
-				if($_q->query()->rowCount() == 0) {
+			->from('SHAREPRICES', array('sum(value) as total'))
+			->where('SHAREPRICES_NAME = ?', $this->_posts['SHAREPRICES_NAME']);
+			$_x = $_q->query()->fetch();
+				//->where('SHAREPRICES_NAME = ?', $this->_posts['SHAREPRICES_NAME']);
+			if(isset($_x['total']) && $_x['total'] == 0) {
 				// Delete
 				$this->_model3->delete(
 						$this->_model3->getAdapter()->quoteInto(
